unread_ops: Centralize window focused logic.
